extracted some components and updated their code
