+++ packages/playground/src/components/index.ts
@@ -1,0 +1,3 @@
+export { CopyLink } from './CopyLink';
+export { GeoPosition } from './GeoPosition';
+export { ThemeSelector } from './ThemeSelector';
